# WIP: Accelerate StringLatin1.indexof (string search version) Intrinsic on Power.