There are different versions of `StringLatin1.indexOf`. This work item is about `java/lang/StringLatin1.indexOf([BI[BII)I` which is used to search for the index of a String inside of another String.