@IBMJimmyk 2025-07-29
I ran `StringIndexOfStringBench` from Bumblebench for performance testing:
https://github.com/adoptium/bumblebench/blob/4ec780b716e7617094cce740d3b085ca4637ee0f/net/adoptopenjdk/bumblebench/string/StringIndexOfStringBench.java

It has three options.
`stringLength` is the length of the sub string to search for.
```
Short = 8-20 characters
Medium = 80-120 characters
Long = 800-1200 characters
Ultra Long = 9000-11000 characters
```

`indexLevel` is the index in the longer string the sub string will be found at.
```
Low = 1-50
Medium = 80-120
High = 800-1200
Ultra High = 48000-51000
```

`useRandomString` is set to use a random distribution of letters for the larger string being searched in. If false, a Wikipedia excerpt is used instead that has a more typical distribution of Latin1 characters foe English text.

Results (higher number is better):
| stringLength | indexLevel | useRandomString | Baseline | Accelerated | Speedup |
| --- | --- | --- | --- | --- | --- |
| Short | Low | N | 37780388 | 65798424 | 1.74x |
| Short | Medium | N | 11950272 | 24247376 | 2.03x |
| Short | High | N | 1260837 | 2548370 | 2.02x |
| Short | Ultra High | N | 13179 | 26197 | 1.99x |
| Medium | Low | N | 37715440 | 54070372 | 1.43x |
| Medium | Medium | N | 11856987 | 22221912 | 1.87x |
| Medium | High | N | 1250761 | 2513439 | 2.01x |
| Medium | Ultra High | N | 11665 | 22283 | 1.91x |
| Long | Low | N | 11449521 | 14981757 | 1.31x |
| Long | Medium | N | 7059095 | 10581082 | 1.50x |
| Long | High | N | 1174269 | 2192076 | 1.87x |
| Long | Ultra High | N | 11500 | 22292 | 1.94x |
| Ultra Long | Low | N | 315025 | 2289270 | 7.27x |
| Ultra Long | Medium | N | 2095266 | 2185265 | 1.04x |
| Ultra Long | High | N | 256952 | 1192369 | 4.64x |
| Ultra Long | Ultra High | N | 11178 | 22100 | 1.98x |
| Short | Low | Y | 37755632 | 93064976 | 2.46x |
| Short | Medium | Y | 20522852 | 69877160 | 3.40x |
| Short | High | Y | 2193947 | 12700037 | 5.79x |
| Short | Ultra High | Y | 47860 | 241128 | 5.04x |
| Medium | Low | Y | 15013809 | 73015424 | 4.86x |
| Medium | Medium | Y | 19562160 | 56882448 | 2.91x |
| Medium | High | Y | 2187883 | 12136565 | 5.55x |
| Medium | Ultra High | Y | 47723 | 241524 | 5.06x |
| Long | Low | Y | 2705928 | 16425695 | 6.07x |
| Long | Medium | Y | 10588714 | 15586048 | 1.47x |
| Long | High | Y | 1994101 | 7730786 | 3.88x |
| Long | Ultra High | Y | 46025 | 238345 | 5.18x |
| Ultra Long | Low | Y | 307301 | 2236618 | 7.28x |
| Ultra Long | Medium | Y | 304372 | 2231100 | 7.33x |
| Ultra Long | High | Y | 270663 | 1940732 | 7.17x |
| Ultra Long | Ultra High | Y | 41331 | 217090 | 5.25x |

Performance has increased across the board. Smallest gain is 1.04x. Largest gain is 7.33x.