@IBMJimmyk 2025-08-25
I have 3 new test configurations. Same as before, `StringLatin1.indexOf` is called 500 million times per test configuration.

These numbers are run on a Power 10 machine but `TR_DisableP10ArrayCmp` is set so the P10 acceleration is not used since the new intrinsic doesn't have a P10 version yet.

The data is in the attached spreadsheet but the quick summary is the new implementation is faster in all cases although some cases are almost a tie.

Test 1:
In this test the main string and the sub string are the same length and contents with one exception.
The character at the index indicated by "Mismatch ID" is different.
This tests the performance of finding a mismatch.
If the mismatch ID is not 0, only the first character of the main string matches the first character of the sub string.
If the mismatch ID is 0, a first character match of the sub string is never found.
Every configuration for string size up to 47 is tested.

Best case: Reduces time by 66%.
Worst case: Reduces time by 25%.

Test 2:
In this test the substring will match the characters in the main string starting at Match Idx.
The first character of the substring will not match any character before Match Idx.
After the first character of the substring matches at Match Idx, the rest of the substring will match as well.
The substring is determined by "Substring Length".
The main string is exactly long enough to hold the sub string at match Idx (so it is "SubstringLength+MatchIdx" long)
This tests matches at index 0-8, 14, 15, 16, 30, 31, 32, 46, 47 and 48
This tests substrings of length 1-64, 128, 256, 512, 1024

Best case: Reduces time by 81%.
Worst case: Reduces time by 3%.

Test 3:
In this test the substring will match the characters in the main string starting at Match Idx+2
The first character of the substring will match the character at Match Idx but the rest of the substring won't be found,
The first character of the substring will also match the character at Match Idx+1 but the rest of the substring still won't be found,
After the first character of the substring matches at Match Idx+2, the rest of the substring will match as well.
The substring is determined by "Substring Length".
The main string is exactly long enough to hold the sub string at match Idx+2 (so it is "SubstringLength+MatchIdx+2" long)
This tests matches at index 0-8, 14, 15, 16, 30, 31, 32, 46, 47 and 48
This tests substrings of length 1-64, 128, 256, 512, 1024

Best case: Reduces time by 74%.
Worst case: Reduces time by 2%.

[IndexOfStrPerf-2.xlsx](https://github.com/user-attachments/files/21975818/IndexOfStrPerf-2.xlsx)