@IBMJimmyk 2025-09-03
I am once again running my test that calls `StringLatin1.indexOf` 500 million times in a loop.

This is run on a P10 machine this time without `TR_DisableP10ArrayCmp` being set.
This compares my new implementation that doesn't have P10 optimizations with the old version that does have P10 optimizations.

The test being run is both the sub string and main string are the same length and have the same contents so there is a match at index 0.
This triggers the Arraycmp acceleration in the original implementation and the P10 version is used due to `TR_DisableP10ArrayCmp` not being set.

| Match Length | Old (s) | New (s) | Ratio |
| --- | --- | --- | --- |
| 1 | 1.14 | 0.46 | 0.40 |
| 2 | 2.01 | 0.66 | 0.33 |
| 3 | 2.30 | 0.65 | 0.28 |
| 4 | 2.30 | 0.95 | 0.41 |
| 5 | 2.29 | 0.90 | 0.39 |
| 6 | 2.30 | 1.02 | 0.45 |
| 7 | 2.29 | 1.02 | 0.45 |
| 8 | 2.30 | 1.02 | 0.45 |
| 9 | 2.29 | 1.02 | 0.45 |
| 10 | 2.29 | 1.27 | 0.56 |
| 11 | 2.29 | 1.27 | 0.56 |
| 12 | 2.30 | 1.40 | 0.61 |
| 13 | 2.29 | 1.40 | 0.61 |
| 14 | 2.29 | 1.53 | 0.67 |
| 15 | 2.30 | 1.52 | 0.66 |
| 16 | 2.29 | 1.15 | 0.50 |
| 17 | 2.52 | 1.15 | 0.45 |
| 18 | 2.53 | 1.40 | 0.55 |
| 19 | 2.52 | 1.40 | 0.55 |
| 20 | 2.53 | 1.65 | 0.65 |
| 21 | 2.53 | 1.65 | 0.65 |
| 22 | 2.53 | 1.78 | 0.70 |
| 23 | 2.53 | 1.77 | 0.70 |
| 24 | 2.53 | 1.65 | 0.65 |
| 25 | 2.53 | 1.65 | 0.65 |
| 26 | 2.53 | 1.90 | 0.75 |
| 27 | 2.53 | 1.90 | 0.75 |
| 28 | 2.52 | 2.02 | 0.80 |
| 29 | 2.52 | 2.02 | 0.80 |
| 30 | 2.52 | 2.15 | 0.85 |
| 31 | 2.53 | 2.15 | 0.85 |
| 32 | 2.53 | 1.40 | 0.55 |
| 33 | 2.76 | 1.40 | 0.51 |
| 34 | 2.76 | 1.65 | 0.60 |
| 35 | 2.76 | 1.65 | 0.60 |
| 36 | 2.76 | 1.90 | 0.69 |
| 37 | 2.76 | 1.90 | 0.69 |
| 38 | 2.76 | 2.03 | 0.73 |
| 39 | 2.76 | 2.03 | 0.73 |
| 40 | 2.76 | 1.91 | 0.69 |
| 41 | 2.77 | 1.90 | 0.69 |
| 42 | 2.76 | 2.16 | 0.78 |
| 43 | 2.76 | 2.16 | 0.78 |
| 44 | 2.76 | 2.28 | 0.83 |
| 45 | 2.76 | 2.28 | 0.83 |
| 46 | 2.77 | 2.41 | 0.87 |
| 47 | 2.76 | 2.43 | 0.88 |
| 48 | 2.76 | 1.68 | 0.61 |
| 49 | 2.99 | 1.68 | 0.56 |
| 50 | 2.99 | 1.90 | 0.64 |
| 51 | 2.99 | 1.90 | 0.64 |
| 52 | 2.99 | 2.15 | 0.72 |
| 53 | 2.99 | 2.15 | 0.72 |
| 54 | 2.99 | 2.28 | 0.76 |
| 55 | 2.99 | 2.28 | 0.76 |
| 56 | 2.99 | 2.15 | 0.72 |
| 57 | 2.99 | 2.15 | 0.72 |
| 58 | 2.99 | 2.41 | 0.81 |
| 59 | 2.99 | 2.41 | 0.81 |
| 60 | 2.99 | 2.53 | 0.85 |
| 61 | 3.00 | 2.53 | 0.84 |
| 62 | 2.99 | 2.66 | 0.89 |
| 63 | 2.99 | 2.66 | 0.89 |
| 64 | 2.99 | 2.15 | 0.72 |
| 128 | 4.90 | 4.13 | 0.84 |
| 256 | 8.29 | 7.60 | 0.92 |
| 512 | 11.27 | 11.65 | 1.03 |
| 1024 | 21.77 | 23.12 | 1.06 |

Performance is still better in almost all cases. The difference in performance for larger arrays is likely a testing quirk since the same 6 instruction loop is used for the bulk of the work performing the array cmp in both cases. So I think performance here is good enough to check in even without P10 optimization that can be added in the future.