@IBMJimmyk 2025-09-03
Same as before, this test calls `StringLatin1.indexOf` 500 million times in a loop.

This time the numbers are for a Power 9 ppcle machine.

The first test being run is the full mismatch case. The first character of the substring can not be found anywhere in the main string.
This focuses on testing performance of searching for a match of the first character of the substring.

| Mismatch Length | Old (s) | New (s) | Ratio |
| --- | --- | --- | --- |
| 1 | 1.08 | 1.05 | 0.97 |
| 2 | 1.95 | 4.23 | 2.17 |
| 3 | 2.25 | 4.23 | 1.88 |
| 4 | 2.60 | 3.10 | 1.20 |
| 5 | 3.31 | 2.56 | 0.77 |
| 6 | 3.43 | 2.56 | 0.75 |
| 7 | 4.34 | 2.56 | 0.59 |
| 8 | 4.94 | 2.56 | 0.52 |
| 9 | 3.37 | 2.56 | 0.76 |
| 10 | 3.74 | 2.56 | 0.68 |
| 11 | 4.00 | 2.56 | 0.64 |
| 12 | 4.32 | 2.56 | 0.59 |
| 13 | 4.66 | 2.56 | 0.55 |
| 14 | 5.03 | 2.56 | 0.51 |
| 15 | 5.42 | 2.56 | 0.47 |
| 16 | 5.78 | 2.31 | 0.40 |
| 17 | 4.93 | 3.17 | 0.64 |
| 18 | 5.25 | 3.17 | 0.60 |
| 19 | 5.55 | 3.17 | 0.57 |
| 20 | 5.92 | 3.17 | 0.53 |
| 21 | 6.26 | 3.17 | 0.51 |
| 22 | 6.64 | 3.17 | 0.48 |
| 23 | 7.01 | 3.17 | 0.45 |
| 24 | 7.39 | 3.17 | 0.43 |
| 25 | 6.56 | 3.17 | 0.48 |
| 26 | 6.87 | 3.17 | 0.46 |
| 27 | 7.14 | 3.17 | 0.44 |
| 28 | 7.48 | 3.17 | 0.42 |
| 29 | 7.83 | 3.17 | 0.40 |
| 30 | 8.19 | 3.17 | 0.39 |
| 31 | 8.56 | 3.17 | 0.37 |
| 32 | 8.94 | 2.77 | 0.31 |
| 33 | 8.19 | 3.50 | 0.43 |
| 34 | 8.43 | 3.50 | 0.42 |
| 35 | 8.71 | 3.50 | 0.40 |
| 36 | 9.05 | 3.50 | 0.39 |
| 37 | 9.41 | 3.50 | 0.37 |
| 38 | 9.78 | 3.50 | 0.36 |
| 39 | 10.14 | 3.50 | 0.35 |
| 40 | 10.52 | 3.50 | 0.33 |
| 41 | 9.87 | 3.50 | 0.36 |
| 42 | 10.22 | 3.50 | 0.34 |
| 43 | 10.47 | 3.50 | 0.33 |
| 44 | 10.85 | 3.50 | 0.32 |
| 45 | 11.19 | 3.50 | 0.31 |
| 46 | 11.55 | 3.50 | 0.30 |
| 47 | 11.92 | 3.50 | 0.29 |
| 48 | 12.36 | 3.12 | 0.25 |
| 49 | 11.46 | 3.81 | 0.33 |
| 50 | 11.80 | 3.82 | 0.32 |
| 51 | 12.10 | 3.81 | 0.32 |
| 52 | 12.46 | 3.81 | 0.31 |
| 53 | 12.82 | 3.81 | 0.30 |
| 54 | 13.16 | 3.81 | 0.29 |
| 55 | 13.54 | 3.81 | 0.28 |
| 56 | 13.95 | 3.81 | 0.27 |
| 57 | 13.05 | 3.81 | 0.29 |
| 58 | 13.38 | 3.81 | 0.28 |
| 59 | 13.86 | 3.81 | 0.28 |
| 60 | 14.21 | 3.81 | 0.27 |
| 61 | 14.51 | 3.81 | 0.26 |
| 62 | 14.97 | 3.81 | 0.25 |
| 63 | 15.82 | 3.81 | 0.24 |
| 64 | 16.28 | 3.48 | 0.21 |
| 128 | 30.51 | 5.55 | 0.18 |
| 256 | 56.94 | 9.27 | 0.16 |
| 512 | 107.54 | 16.84 | 0.16 |
| 1024 | 208.59 | 35.25 | 0.17 |

My accelerated assembly has a special case to fast path a check of the first character of the substring with the first character of the main string so performance is as good as the original implementation which is naturally good with 1 byte long main strings. Performance isn't as good for 2, 3 and 4 byte long main string as byte by byte comparison still beats the vector load in performance. For 5 byte strings, the vector implementation is now better and becomes much better for longer strings.

This is once again the full match case where the sub string and main string have the same contents and will match at index 0. It is very similar to an arraycmp.

| Match Length | Old (s) | New (s) | Ratio |
| --- | --- | --- | --- |
| 1 | 1.20 | 0.51 | 0.43 |
| 2 | 2.42 | 0.80 | 0.33 |
| 3 | 2.98 | 0.80 | 0.27 |
| 4 | 3.27 | 0.98 | 0.30 |
| 5 | 2.65 | 0.98 | 0.37 |
| 6 | 2.98 | 1.11 | 0.37 |
| 7 | 3.31 | 1.11 | 0.33 |
| 8 | 3.61 | 1.10 | 0.31 |
| 9 | 2.87 | 1.13 | 0.39 |
| 10 | 3.29 | 1.36 | 0.41 |
| 11 | 3.62 | 1.41 | 0.39 |
| 12 | 3.90 | 1.50 | 0.39 |
| 13 | 3.29 | 1.54 | 0.47 |
| 14 | 3.62 | 1.64 | 0.45 |
| 15 | 3.92 | 1.69 | 0.43 |
| 16 | 4.19 | 1.64 | 0.39 |
| 17 | 3.34 | 1.64 | 0.49 |
| 18 | 3.76 | 1.84 | 0.49 |
| 19 | 4.03 | 1.91 | 0.47 |
| 20 | 4.28 | 2.12 | 0.49 |
| 21 | 3.85 | 2.19 | 0.57 |
| 22 | 4.11 | 2.25 | 0.55 |
| 23 | 4.39 | 2.33 | 0.53 |
| 24 | 4.62 | 2.12 | 0.46 |
| 25 | 3.92 | 2.19 | 0.56 |
| 26 | 4.28 | 2.40 | 0.56 |
| 27 | 4.47 | 2.47 | 0.55 |
| 28 | 4.73 | 2.53 | 0.54 |
| 29 | 4.28 | 2.61 | 0.61 |
| 30 | 4.48 | 2.67 | 0.60 |
| 31 | 4.77 | 2.75 | 0.58 |
| 32 | 5.00 | 2.12 | 0.42 |
| 33 | 3.86 | 2.19 | 0.57 |
| 34 | 4.29 | 2.33 | 0.54 |
| 35 | 4.57 | 2.49 | 0.54 |
| 36 | 4.79 | 2.62 | 0.55 |
| 37 | 4.39 | 2.83 | 0.64 |
| 38 | 4.65 | 2.74 | 0.59 |
| 39 | 4.92 | 2.98 | 0.61 |
| 40 | 5.13 | 2.62 | 0.51 |
| 41 | 4.40 | 2.85 | 0.65 |
| 42 | 4.76 | 3.00 | 0.63 |
| 43 | 5.01 | 3.12 | 0.62 |
| 44 | 5.25 | 3.16 | 0.60 |
| 45 | 4.76 | 3.28 | 0.69 |
| 46 | 5.01 | 3.34 | 0.67 |
| 47 | 5.31 | 3.44 | 0.65 |
| 48 | 5.55 | 2.66 | 0.48 |
| 49 | 4.34 | 2.77 | 0.64 |
| 50 | 4.72 | 2.90 | 0.61 |
| 51 | 5.00 | 3.16 | 0.63 |
| 52 | 5.22 | 3.12 | 0.60 |
| 53 | 4.81 | 3.44 | 0.72 |
| 54 | 5.08 | 3.26 | 0.64 |
| 55 | 5.35 | 3.57 | 0.67 |
| 56 | 5.57 | 3.12 | 0.56 |
| 57 | 4.81 | 3.43 | 0.71 |
| 58 | 5.17 | 3.53 | 0.68 |
| 59 | 5.44 | 3.69 | 0.68 |
| 60 | 5.68 | 3.71 | 0.65 |
| 61 | 5.17 | 3.86 | 0.75 |
| 62 | 5.44 | 3.92 | 0.72 |
| 63 | 5.74 | 4.06 | 0.71 |
| 64 | 5.99 | 3.24 | 0.54 |
| 128 | 7.73 | 5.74 | 0.74 |
| 256 | 11.28 | 10.99 | 0.97 |
| 512 | 18.48 | 21.46 | 1.16 |
| 1024 | 34.97 | 43.88 | 1.25 |

Once again, performance is good for shorter strings but seemingly not quite as good for longer strings. Once again, I think this is due to a testing quirk since both versions will run the same instructions as a vector loop to perform most of the arraycmp for longer arrays.