@IBMJimmyk 2025-09-03
These results are for the same tests but for AIX on P9. Overall they are similar but slightly better than the ppcle P9 results.
Only the 2 and 3 length mismatched strings lengths are slower and not by as much.
For matching strings, performance is always better.

| Mismatch Length | Old (s) | New (s) | Ratio |
| --- | --- | --- | --- |
| 1 | 1.08 | 0.66 | 0.61 |
| 2 | 1.96 | 2.56 | 1.31 |
| 3 | 2.26 | 2.56 | 1.13 |
| 4 | 2.60 | 2.56 | 0.98 |
| 5 | 3.29 | 2.56 | 0.78 |
| 6 | 3.44 | 2.56 | 0.75 |
| 7 | 4.35 | 2.56 | 0.59 |
| 8 | 4.95 | 2.56 | 0.52 |
| 9 | 3.38 | 2.56 | 0.76 |
| 10 | 3.75 | 2.56 | 0.68 |
| 11 | 4.00 | 2.56 | 0.64 |
| 12 | 4.33 | 2.56 | 0.59 |
| 13 | 4.67 | 2.56 | 0.55 |
| 14 | 5.03 | 2.56 | 0.51 |
| 15 | 5.43 | 2.56 | 0.47 |
| 16 | 5.80 | 2.31 | 0.40 |
| 17 | 4.94 | 3.18 | 0.64 |
| 18 | 5.25 | 3.17 | 0.60 |
| 19 | 5.54 | 3.17 | 0.57 |
| 20 | 5.93 | 3.17 | 0.54 |
| 21 | 6.28 | 3.17 | 0.51 |
| 22 | 6.65 | 3.17 | 0.48 |
| 23 | 7.02 | 3.17 | 0.45 |
| 24 | 7.40 | 3.17 | 0.43 |
| 25 | 6.56 | 3.17 | 0.48 |
| 26 | 6.88 | 3.17 | 0.46 |
| 27 | 7.15 | 3.17 | 0.44 |
| 28 | 7.49 | 3.17 | 0.42 |
| 29 | 7.84 | 3.17 | 0.40 |
| 30 | 8.20 | 3.17 | 0.39 |
| 31 | 8.57 | 3.17 | 0.37 |
| 32 | 8.95 | 2.78 | 0.31 |
| 33 | 8.15 | 3.51 | 0.43 |
| 34 | 8.44 | 3.51 | 0.42 |
| 35 | 8.72 | 3.51 | 0.40 |
| 36 | 9.06 | 3.51 | 0.39 |
| 37 | 9.41 | 3.51 | 0.37 |
| 38 | 9.79 | 3.51 | 0.36 |
| 39 | 10.15 | 3.51 | 0.35 |
| 40 | 10.53 | 3.51 | 0.33 |
| 41 | 9.88 | 3.51 | 0.36 |
| 42 | 10.22 | 3.51 | 0.34 |
| 43 | 10.48 | 3.51 | 0.33 |
| 44 | 10.87 | 3.51 | 0.32 |
| 45 | 11.19 | 3.51 | 0.31 |
| 46 | 11.56 | 3.51 | 0.30 |
| 47 | 11.94 | 3.51 | 0.29 |
| 48 | 12.37 | 3.12 | 0.25 |
| 49 | 11.46 | 3.82 | 0.33 |
| 50 | 11.78 | 3.82 | 0.32 |
| 51 | 12.14 | 3.82 | 0.31 |
| 52 | 12.44 | 3.82 | 0.31 |
| 53 | 12.83 | 3.82 | 0.30 |
| 54 | 14.76 | 3.82 | 0.26 |
| 55 | 13.53 | 3.82 | 0.28 |
| 56 | 13.95 | 3.82 | 0.27 |
| 57 | 13.05 | 3.82 | 0.29 |
| 58 | 13.38 | 3.82 | 0.29 |
| 59 | 13.96 | 3.82 | 0.27 |
| 60 | 14.11 | 3.82 | 0.27 |
| 61 | 14.47 | 3.82 | 0.26 |
| 62 | 14.81 | 3.82 | 0.26 |
| 63 | 15.29 | 3.82 | 0.25 |
| 64 | 15.84 | 3.49 | 0.22 |
| 128 | 30.46 | 5.55 | 0.18 |
| 256 | 56.99 | 9.22 | 0.16 |
| 512 | 107.58 | 16.79 | 0.16 |
| 1024 | 210.30 | 33.27 | 0.16 |

| Match Length | Old (s) | New (s) | Ratio |
| --- | --- | --- | --- |
| 1 | 1.20 | 0.64 | 0.53 |
| 2 | 2.28 | 0.98 | 0.43 |
| 3 | 2.99 | 0.97 | 0.33 |
| 4 | 3.27 | 0.99 | 0.30 |
| 5 | 2.65 | 0.99 | 0.37 |
| 6 | 2.99 | 1.11 | 0.37 |
| 7 | 3.30 | 1.11 | 0.34 |
| 8 | 3.62 | 1.11 | 0.31 |
| 9 | 2.88 | 1.14 | 0.40 |
| 10 | 3.30 | 1.50 | 0.46 |
| 11 | 3.62 | 1.51 | 0.42 |
| 12 | 3.91 | 1.51 | 0.39 |
| 13 | 3.27 | 1.54 | 0.47 |
| 14 | 3.62 | 1.64 | 0.45 |
| 15 | 3.92 | 1.70 | 0.43 |
| 16 | 4.19 | 1.48 | 0.35 |
| 17 | 3.38 | 1.45 | 0.43 |
| 18 | 3.77 | 1.70 | 0.45 |
| 19 | 4.04 | 1.72 | 0.42 |
| 20 | 4.29 | 1.95 | 0.45 |
| 21 | 3.83 | 1.98 | 0.52 |
| 22 | 4.15 | 2.08 | 0.50 |
| 23 | 4.39 | 2.11 | 0.48 |
| 24 | 4.62 | 1.95 | 0.42 |
| 25 | 3.93 | 1.99 | 0.51 |
| 26 | 4.28 | 2.22 | 0.52 |
| 27 | 4.50 | 2.26 | 0.50 |
| 28 | 4.76 | 2.34 | 0.49 |
| 29 | 4.28 | 2.39 | 0.56 |
| 30 | 4.50 | 2.47 | 0.55 |
| 31 | 4.78 | 2.52 | 0.53 |
| 32 | 5.01 | 1.85 | 0.37 |
| 33 | 3.87 | 1.88 | 0.49 |
| 34 | 4.30 | 2.05 | 0.48 |
| 35 | 4.57 | 2.18 | 0.48 |
| 36 | 4.79 | 2.33 | 0.49 |
| 37 | 4.40 | 2.47 | 0.56 |
| 38 | 4.67 | 2.44 | 0.52 |
| 39 | 4.92 | 2.64 | 0.54 |
| 40 | 5.14 | 2.32 | 0.45 |
| 41 | 4.40 | 2.49 | 0.57 |
| 42 | 4.77 | 2.67 | 0.56 |
| 43 | 5.02 | 2.77 | 0.55 |
| 44 | 5.26 | 2.81 | 0.53 |
| 45 | 4.77 | 2.91 | 0.61 |
| 46 | 5.02 | 2.94 | 0.59 |
| 47 | 5.31 | 3.03 | 0.57 |
| 48 | 5.55 | 2.25 | 0.41 |
| 49 | 4.35 | 2.24 | 0.51 |
| 50 | 4.74 | 2.42 | 0.51 |
| 51 | 5.01 | 2.53 | 0.50 |
| 52 | 5.22 | 2.84 | 0.54 |
| 53 | 4.82 | 2.97 | 0.62 |
| 54 | 5.10 | 2.94 | 0.58 |
| 55 | 5.36 | 3.15 | 0.59 |
| 56 | 5.58 | 2.70 | 0.48 |
| 57 | 4.82 | 2.89 | 0.60 |
| 58 | 5.17 | 3.09 | 0.60 |
| 59 | 5.46 | 3.17 | 0.58 |
| 60 | 5.68 | 3.21 | 0.57 |
| 61 | 5.17 | 3.29 | 0.64 |
| 62 | 5.46 | 3.35 | 0.61 |
| 63 | 5.75 | 3.42 | 0.59 |
| 64 | 6.00 | 2.64 | 0.44 |
| 128 | 7.74 | 4.22 | 0.55 |
| 256 | 11.30 | 7.53 | 0.67 |
| 512 | 18.49 | 14.63 | 0.79 |
| 1024 | 35.31 | 29.29 | 0.83 |